+++ src/main/java/nom/bdezonia/zorbage/type/data/float16/real/Float16VectorMember.java
@@ -649,7 +649,7 @@ public final class Float16VectorMember
 		if (component == 0) {
 			Float16Member tmp = tmpFloat.get();
 			v(index.get(0), tmp);
-			return BigInteger.valueOf((long) tmp.v());
+			return BigDecimal.valueOf(tmp.v()).toBigInteger();
 		}
 		return BigInteger.ZERO;
 	}
@@ -873,7 +873,7 @@ public final class Float16VectorMember
 		else {
 			Float16Member tmp = tmpFloat.get();
 			v(index.get(0), tmp);
-			return BigInteger.valueOf((long) tmp.v());
+			return BigDecimal.valueOf(tmp.v()).toBigInteger();
 		}
 	}
